+++ config/test.exs
@@ -21,6 +21,7 @@ config :td_ie, TdIe.Repo,
   pool_size: 1
 
 config :td_core, TdCore.Search.Cluster, api: ElasticsearchMock
+config :td_core, TdCore.Search.IndexWorker, TdCore.Search.IndexWorkerMock
 
 config :td_cache, :audit, stream: "audit:events:test"
 config :td_cache, redis_host: "redis", port: 6380