+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- [TD-6561] Standardise aggregations limits 
+
 ## [6.5.0] 2024-04-30
 
 ### Added